backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
numpy==1.24.3
scipy==1.11.2
pandas==2.0.3
APScheduler==3.10.4
python-dotenv==1.0.0
pytz==2023.3
yfinance>=0.2.30
finvizfinance>=1.3.0
google-genai>=1.74.0
